move SUPPORTED_FUNCTIONS to query builder
