+++ .changeset/heavy-clocks-sniff.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-studio-preset-query-builder': patch
+---
+
+**BREAKING CHANGE:** Rename `@finos/legend-studio-plugin-query-builder` to `@finos/legend-studio-preset-query-builder`. `SUPPORTED_FUNCTIONS` is no-longer exported from core, but is not part of query builder preset.
