+++ packages/legend-studio-preset-query-builder/src/QueryBuilder_Constants.ts
@@ -22,4 +22,24 @@ export enum QUERY_BUILDER_TEST_ID {
   QUERY_BUILDER_FILTER = 'query__builder__filter',
 }
 
+export enum SUPPORTED_FUNCTIONS {
+  TDS_ASC = 'asc', // meta::pure::tds::asc
+  TDS_DESC = 'desc', // meta::pure::tds::desc
+  TDS_AGG = 'agg', // meta::pure::tds::agg
+  TDS_COL = 'col', // meta::pure::tds::col
+  TDS_DISTINCT = 'distinct', // meta::pure::tds::distinct
+  TDS_PROJECT = 'project', // meta::pure::tds::project
+  TDS_GROUP_BY = 'groupBy', // meta::pure::tds::groupBy
+  TDS_SORT = 'sort', // meta::pure::tds::sort
+  TDS_TAKE = 'take', // meta::pure::tds::take
+  EXISTS = 'exists', // meta::pure::functions::collection::exists
+  FILTER = 'filter', // meta::pure::functions::collection::filter
+  GET_ALL = 'getAll', // meta::pure::functions::collection::getAll
+  GRAPH_FETCH = 'graphFetch', // meta::pure::graphFetch::execution::graphFetch
+  GRAPH_FETCH_CHECKED = 'graphFetchChecked', // meta::pure::graphFetch::execution::graphFetchChecked
+  SERIALIZE = 'serialize', // meta::pure::graphFetch::execution::serialize
+  NOT = 'not', // meta::pure::functions::boolean::not
+  TAKE = 'take', // meta::pure::functions::collection::take
+}
+
 export const DEFAULT_LAMBDA_VARIABLE_NAME = 'x';
