+++ packages/legend-studio-preset-query-builder/src/components/__tests__/QueryBuilderLambdaProcesser.test.tsx
@@ -42,9 +42,11 @@ import {
   AbstractPropertyExpression,
   RawLambda,
   setUpEditorWithDefaultSDLCData,
-  SUPPORTED_FUNCTIONS,
 } from '@finos/legend-studio';
-import { QUERY_BUILDER_TEST_ID } from '../../QueryBuilder_Constants';
+import {
+  QUERY_BUILDER_TEST_ID,
+  SUPPORTED_FUNCTIONS,
+} from '../../QueryBuilder_Constants';
 import { QueryBuilderState } from '../../stores/QueryBuilderState';
 import { flowResult } from 'mobx';
 import { buildQueryBuilderMockedEditorStore } from './QueryBuilder_TestUtils';
