+++ packages/legend-studio/src/index.ts
@@ -62,7 +62,6 @@ export { VariableExpression } from './models/metamodels/pure/model/valueSpecific
 export {
   AbstractPropertyExpression,
   SimpleFunctionExpression,
-  SUPPORTED_FUNCTIONS,
   FunctionExpression,
 } from './models/metamodels/pure/model/valueSpecification/SimpleFunctionExpression';
 export {
